% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{inplace-op}
\alias{inplace-op}
\alias{\%*<-\%}
\alias{\%+<-\%}
\title{In-place operators}
\usage{
x \%*<-\% value

x \%+<-\% value
}
\arguments{
\item{x}{Variable.}

\item{value}{Value.}
}
\value{
\code{NULL} (invisibly).
}
\description{
In-place operators
}
