% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{inplace-sweep}
\alias{inplace-sweep}
\alias{sweep1_in_place}
\alias{sweep2_in_place}
\title{Sweep}
\usage{
sweep1_in_place(x, STATS, FUN = c("-", "+", "*", "/"))

sweep2_in_place(x, STATS, FUN = c("-", "+", "*", "/"))
}
\arguments{
\item{x}{a matrix.}

\item{STATS}{the summary statistic which is to be swept out.}

\item{FUN}{the function to be used to carry out the sweep.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Sweep
}
\examples{
x <- matrix(1:6, 2)
sweep(x, 2, 1:3)
x
sweep2_in_place(x, 1:3)
x

}
