% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace-package.R
\docType{package}
\name{inplace-package}
\alias{inplace}
\alias{inplace-package}
\title{inplace: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
