% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inplace-package.R
\docType{package}
\name{inplace-package}
\alias{inplace}
\alias{inplace-package}
\title{inplace: In-place Operators for R}
\description{
It provides in-place operators for R
that are equivalent to '+=', '-=', '*=', '/=' in C++.
Those can be applied on integer|double vectors|matrices.
}
\author{
\strong{Maintainer}: Florian Privé \email{florian.prive.21@gmail.com}

}
\keyword{internal}
